First push
all the changes done offline

+++ lib/i2cdev/I2Cdev.cpp
@@ -278,7 +278,7 @@ int8_t I2Cdev::readBytes(uint8_t devAddr, uint8_t regAddr, uint8_t length, uint8
             // I2C/TWI subsystem uses internal buffer that breaks with large data requests
             // so if user requests more than BUFFER_LENGTH bytes, we have to do it in
             // smaller chunks instead of all at once
-            for (uint8_t k = 0; k < length; k += min(length, BUFFER_LENGTH)) {
+            for (uint16_t k = 0; k < length; k += min(length, BUFFER_LENGTH)) {
                 Wire.beginTransmission(devAddr);
                 Wire.write(regAddr);
                 Wire.endTransmission();
@@ -291,10 +291,19 @@ int8_t I2Cdev::readBytes(uint8_t devAddr, uint8_t regAddr, uint8_t length, uint8
                         Serial.print(data[count], HEX);
                         if (count + 1 < length) Serial.print(" ");
                     #endif
+                    // if(length == 140 || length == 56){
+                    //     Serial.print(data[count], HEX);
+                    //     if(count + 1 < length) Serial.print(" ");
+                    //     else Serial.println("");
+                    // }
                 }
 
+                // if(length==140) Serial.printf("begin");
                 status = Wire.endTransmission();
+                // if(length==140) Serial.println(" end");
+                // if(length==140) Serial.println(k);
             }
+            // if(length==140 ||length == 56)Serial.println("exitloop");
         #endif
 
     #elif (I2CDEV_IMPLEMENTATION == I2CDEV_BUILTIN_FASTWIRE)
